feat(catalog): consolidate build pipeline and server-side catalog merge
- Single build step writes dist/catalogs/*.json (drop .catalog-manifest.json)
- CLI tarballs vendor dist/catalogs for lobu run cold-start
- Add include=catalog on installed routes and include_catalog on manage_catalog
- Extract buildCatalogListResponse, merge helpers, move connector-definitions
- Rename operations/catalog.ts to connector-operations.ts
- Bump owletto to use server-side merge for connectors and skills
